# packages/contracts/openapi/v1.yaml
...
openapi: 3.1.0
info:
  description: |-
    Version 1 of the Chaya 02 HTTP API. Breaking changes require a new /api/v2 document.

    Generated from the backend's controllers; packages/contracts/openapi/v1.yaml is the committed output and OpenApiContractTest fails when it drifts.

    Roles (x-chaya-roles) under /api/v1/venues/{venueId}/ are the roles the caller holds at that venue. A resource outside the caller's organization or venue answers 404, never 403.

    Errors are RFC 9457 problems with a stable machine-readable `code`.
  title: Chaya 02 API
  version: v1
servers:
- description: The API's own origin
  url: /
paths:
...
  /api/v1/venues/{venueId}:
    delete:
      description: "Every capture, version, artifact, POI, anchor, floor, search query\
        \ and public link; the venue is 404 afterwards (docs/privacy-erasure.md).\
        \ Organization admin only. Idempotent."
      operationId: erasureEraseVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErasureView"
          description: "Erasure record, COMPLETED"
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErasureView"
          description: "Erasure record, OBJECTS_PENDING"
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: ERASURE_NOT_PERMITTED
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Erase the venue
      tags:
      - erasure-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
    get:
      operationId: venueGet
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: One venue
      tags:
      - venue-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
    patch:
      operationId: venueUpdate
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenue"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: "Rename a venue or change its timezone (admin, venue-manager)"
      tags:
      - venue-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
...
  /api/v1/venues/{venueId}/jobs/{jobId}/cancel:
    post:
      operationId: scanCancel
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: INVALID_JOB_TRANSITION
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Cancel a job
      tags:
      - scan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/jobs/{jobId}/retry:
    post:
      operationId: scanRetry
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: INVALID_JOB_TRANSITION
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Retry a failed job
      tags:
      - scan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
...
  /api/v1/venues/{venueId}/scans:
    post:
      operationId: scanCreateScan
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScan"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedScan"
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Create a scan (legacy job API)
      tags:
      - scan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/scans/{scanId}/jobs:
    post:
      operationId: scanEnqueue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: scanId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnqueueJob"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  format: uuid
          description: Accepted
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Enqueue one processing stage for a scan (legacy job API)
      tags:
      - scan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
...
components:
  schemas:
...
    CreateScan:
      type: object
      properties:
        floorId:
          type: string
          format: uuid
...
    CreatedScan:
      type: object
      properties:
        captureSessionId:
          type: string
          format: uuid
        scanId:
          type: string
          format: uuid
...
    EnqueueJob:
      type: object
      properties:
        stage:
          type: string
          enum:
          - INPUT_VALIDATION
          - FFMPEG_PREPROCESS
          - FRAME_QUALITY_FILTER
          - PRIVACY_PREPROCESS
          - POSE_ESTIMATION
          - SPLAT_RECONSTRUCTION
          - SEMANTIC_SEGMENTATION
          - GEOMETRIC_CLEANUP
          - REGION_ALIGNMENT
          - REGION_SPLICE
          - PLANE_FITTING
          - ARTIFACT_GENERATION
          - NAVIGATION_BAKING
          - SEMANTIC_INDEXING
          - MEDIA_FILTER
          - SPLAT_TRAINING
          - GEOMETRY_CLEANUP
          - SEGMENTATION
          - OBJECT_DETECTION
          - NAVMESH_GENERATION
      required:
      - stage
...
    ErasureView:
      type: object
      properties:
        cascade:
          type: boolean
        completedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        lastError:
          type: string
        objectAttempts:
          type: integer
          format: int32
        requestedAt:
          type: string
          format: date-time
        requestedBy:
          type: string
        rowsErasedAt:
          type: string
          format: date-time
        status:
          type: string
        summary:
          type: object
          additionalProperties: {}
        targetId:
          type: string
          format: uuid
        targetType:
          type: string
          enum:
          - VENUE
          - CAPTURE
          - SCAN_VERSION
        venueId:
          type: string
          format: uuid
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        code:
          type: string
          description: "Stable machine-readable error code, e.g. NOT_FOUND, BAD_REQUEST,\
            \ VERSION_INCOMPLETE. Absent on Spring's own request-parsing and validation\
            \ problems."
        detail:
          type: string
        instance:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
          example: about:blank
      required:
      - status
...
    UpdateVenue:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
...
    Venue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        organizationId:
          type: string
          format: uuid
        slug:
          type: string
        timezone:
          type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Keycloak access token for the chaya-api audience. Organization,\
        \ roles and venue memberships come from its claims."
      scheme: bearer
      type: http
    viewerToken:
      description: "Short-lived public-viewer token from POST /api/v1/public/viewer-token,\
        \ bound to one venue; acts as PUBLIC_VIEWER. Never together with a bearer\
        \ token (400)."
      in: header
      name: X-Chaya-Viewer-Token
      type: apiKey